int rpm -> float rpm

+++ firmware/controllers/algo/launch_control.h
@@ -22,8 +22,8 @@ public:
 	bool isInsideSpeedCondition() const;
 	bool isInsideTpsCondition() const;
 	bool isInsideSwitchCondition();
-	bool isInsideRPMCondition(int rpm) const;
-	bool isLaunchConditionMet(int rpm);
+	bool isInsideRPMCondition(float rpm) const;
+	bool isLaunchConditionMet(float rpm);
 
 	bool isLaunchSparkRpmRetardCondition() const;
 	bool isLaunchFuelRpmRetardCondition() const;
